Move utilLabels to utils/labels
Caleydo/cohort#693
