+++ src/CohortRepresentations.ts
@@ -5,7 +5,7 @@ import { IElement, IRectCohortRep } from './CohortInterfaces';
 import { getRootCohort } from './cohortview';
 import { log } from './util';
 import { CohortRemoveEvent, CohortSelectionEvent } from './base/events';
-import { labelFromFilter } from './utilLabels';
+import { labelFromFilter } from './utils/labels';
 
 export class RectCohortRep implements IRectCohortRep {
   id: string;
