+++ src/app/CoralApp.ts
@@ -29,7 +29,7 @@ import welcomeHtml from './templates/Welcome.html';
 import { getAnimatedLoadingText, handleDataLoadError, log } from '../util';
 import { CohortSelectionEvent, ConfirmTaskEvent, CONFIRM_TASK_EVENT_TYPE, PreviewConfirmEvent } from '../base/events';
 import { idCellline, idCovid19, idStudent, idTissue, IEntitySourceConfig } from '../config/entities';
-import { niceName } from '../utilLabels';
+import { niceName } from '../utils/labels';
 import { IDatasetDesc, IPanelDesc } from './interfaces';
 import { CohortSelectionListener } from './CoralSelectionListener';
